+++ src/tests/unit/tile.unit.test.ts
@@ -1,4 +1,4 @@
-import { Board } from "../logic/Board";
+import { Board } from "../../logic/Board";
 
 let board: Board;
